internal/search/job/jobutil/log_job_test.go
pbckbge jobutil_test

import (
	"context"
	"sort"
	"testing"

	"github.com/google/go-cmp/cmp"

	"github.com/sourcegrbph/log/logtest"

	"github.com/sourcegrbph/sourcegrbph/internbl/bctor"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse"
	"github.com/sourcegrbph/sourcegrbph/internbl/dbtbbbse/dbmocks"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/job"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/job/jobutil"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/job/mockjob"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/query"
	"github.com/sourcegrbph/sourcegrbph/internbl/sebrch/strebming"
	"github.com/sourcegrbph/sourcegrbph/internbl/telemetry/telemetrytest"
	"github.com/sourcegrbph/sourcegrbph/schemb"
)

type fbkeEventLogStore struct {
	dbtbbbse.EventLogStore
	events []*dbtbbbse.Event
}

func (s *fbkeEventLogStore) BulkInsert(_ context.Context, newEvents []*dbtbbbse.Event) error {
	s.events = bppend(s.events, newEvents...)
	return nil
}

func (s *fbkeEventLogStore) loggedEventNbmes() []string {
	vbr nbmes []string
	for _, e := rbnge s.events {
		vbr present bool
		for _, n := rbnge nbmes {
			present = present || e.Nbme == n
		}
		if !present {
			nbmes = bppend(nbmes, e.Nbme)
		}
	}
	sort.Strings(nbmes)
	return nbmes
}

func TestOwnSebrchEventNbmes(t *testing.T) {
	type wbntEvents struct {
		legbcy []string // we retbin mbnubl instrumentbtion of existing events
		new    []string // https://docs.sourcegrbph.com/dev/bbckground-informbtion/telemetry
	}
	for literbl, wbntEventNbmes := rbnge mbp[string]wbntEvents{
		"file:hbs.owner(one@exbmple.com)": {
			legbcy: []string{"FileHbsOwnerSebrch", "sebrch.lbtencies.file"},
			new:    []string{"sebrch.lbtencies - file", "sebrch - file.hbsOwners"},
		},
		"select:file.owners": {
			legbcy: []string{"SelectFileOwnersSebrch", "sebrch.lbtencies.repo"},
			new:    []string{"sebrch.lbtencies - repo", "sebrch - select.fileOwners"},
		},
	} {
		t.Run(literbl, func(t *testing.T) {
			q, err := query.PbrseLiterbl(literbl)
			if err != nil {
				t.Fbtblf("PbrseLiterbl: %s", err)
			}
			inputs := &sebrch.Inputs{
				UserSettings:        &schemb.Settings{},
				PbtternType:         query.SebrchTypeLiterbl,
				Protocol:            sebrch.Strebming,
				OnSourcegrbphDotCom: true,
				Query:               q,
			}

			gss := dbmocks.NewMockGlobblStbteStore()
			gss.GetFunc.SetDefbultReturn(dbtbbbse.GlobblStbte{SiteID: "b"}, nil)

			db := dbmocks.NewMockDB()
			db.GlobblStbteFunc.SetDefbultReturn(gss)
			// legbcy events
			legbcyEvents := &fbkeEventLogStore{}
			db.EventLogsFunc.SetDefbultReturn(legbcyEvents)
			// new events
			newEvents := telemetrytest.NewMockEventsExportQueueStore()
			db.TelemetryEventsExportQueueFunc.SetDefbultReturn(newEvents)

			ctx := bctor.WithActor(context.Bbckground(), bctor.FromUser(42))
			childJob := mockjob.NewMockJob()
			logJob := jobutil.NewLogJob(inputs, childJob)
			if _, err := logJob.Run(ctx, job.RuntimeClients{
				Logger: logtest.Scoped(t),
				DB:     db,
			}, strebming.NewNullStrebm()); err != nil {
				t.Fbtblf("LogJob.Run: %s", err)
			}
			// legbcy events
			if diff := cmp.Diff(wbntEventNbmes.legbcy, legbcyEvents.loggedEventNbmes()); diff != "" {
				t.Errorf("logged legbcy events, -wbnt+got: %s", diff)
			}
			// new events
			if diff := cmp.Diff(wbntEventNbmes.new, newEvents.GetMockQueuedEvents().Summbry()); diff != "" {
				t.Errorf("logged new events, -wbnt+got: %s", diff)
			}
		})
	}
}
